add const modifier to many functions fix several compiler warnings

+++ Platforms/Ogre/OgrePlatform/src/MyGUI_OgreRenderManager.cpp
@@ -143,7 +143,7 @@ namespace MyGUI
 		}
 	}
 
-	void OgreRenderManager::setActiveViewport(size_t _num)
+	void OgreRenderManager::setActiveViewport(unsigned short _num)
 	{
 		mActiveViewport = _num;
